Add different files for drawer components
