+++ src/components/Drawer/index.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export { default as default } from './Drawer';
